reorganize folder structure. move to internal
