+++ app/internal/ranking/quality.py
