+++ app/routers/settings.py
@@ -6,9 +6,11 @@ from aiohttp import ClientResponseError, ClientSession
 from fastapi import APIRouter, Depends, Form, HTTPException, Request, Response
 from sqlmodel import Session, select
 
-from app.db import get_session
+from app.internal.models import EventEnum, GroupEnum, Notification, User
 from app.internal.prowlarr.indexer_categories import indexer_categories
-from app.models import EventEnum, GroupEnum, Notification, User
+from app.internal.prowlarr.notifications import send_notification
+from app.internal.prowlarr.prowlarr import flush_prowlarr_cache, prowlarr_config
+from app.internal.ranking.quality import IndexerFlag, QualityRange, quality_config
 from app.util.auth import (
     DetailedUser,
     LoginTypeEnum,
@@ -19,9 +21,7 @@ from app.util.auth import (
     raise_for_invalid_password,
 )
 from app.util.connection import get_connection
-from app.util.notifications import send_notification
-from app.util.prowlarr import flush_prowlarr_cache, prowlarr_config
-from app.util.ranking.quality import IndexerFlag, QualityRange, quality_config
+from app.util.db import get_session
 from app.util.templates import template_response
 
 router = APIRouter(prefix="/settings")
@@ -578,7 +578,6 @@ def delete_notification(
     notifications = session.exec(select(Notification)).all()
     for notif in notifications:
         if notif.id == notification_id:
-            print("DELETED")
             session.delete(notif)
             session.commit()
             break
