+++ app/util/db.py
@@ -1,4 +1,5 @@
 from contextlib import contextmanager
+
 from sqlalchemy import create_engine
 from sqlmodel import Session, text
